+++ backend/src/main/java/com/car_backend/dto/jobCard/JobCardItemDto.java
@@ -1,5 +1,7 @@
 package com.car_backend.dto.jobCard;
 
+import java.math.BigDecimal;
+
 import lombok.Builder;
 import lombok.Data;
 
@@ -8,8 +10,8 @@ import lombok.Data;
 public class JobCardItemDto {
 	private Long id;
 	private String itemName;
-	private Double itemPrice;
+	private BigDecimal itemPrice;
 	private Integer quantity;
-	private Double totalPrice;
+	private BigDecimal totalPrice;
 	
 }
